Now you can set your own credentials :D
Finally, honestly it was hard to get it done @_@

+++ Login_Script/login.sh
@@ -1,32 +1,67 @@
 #! /bin/bash
-#Simple login script which asks for the predefined usrnm & pswd
-#it terminates the app if credentials dont match.
+#Login script which asks for the usrnm & pswd which the user defines, it terminates the app if credentials dont match.
 #It has improved now, but still not totally foolproof. 
 #Version: 0.3
 
+#Path where the valuables are stored.
+pt=../usr/share/.login
+
+#Function that sets credentials.
+st_crd() {
+	read -p "Set Your Username: " un
+	read -s -p "Set Your Password: " pf
+
+	if [[ $un != "" && $pf != "" ]];
+	then
+		touch $pt
+	        echo -e "$un\n$pf" >> $pt
+	        echo -e "\nCredentials added!!"
+		echo -e "\nExiting Termux, please restart it!"
+		sleep 3
+		kill -9 $PPID
+	else
+		echo -e '\nCredentials cannot be empty!'
+		exit
+	fi
+
+}
+
+#Trap the intruder ;)
 int_trp() {
-	echo "Interrupted"
+	echo "Intruder Alert!!"
 	sleep 3
 	kill -9 $PPID
 }
 
-trap 'int_trp' INT
+#Function for checking if the user is authorized.
+chk_crd() {
+	trap 'int_trp' INT
 
-echo "Username"
-read n
-echo "Password"
-read -s p
+	unm=$( head -n 1 $pt )
+	p=$( tail -n 1 $pt )
 
-#enter your credentials here
-if [[ $n == "Ali" && $p == "Sars" ]];
+	echo "Enter Un"
+	read n
+	echo "Enter Passwd"
+	read -s pss
+
+	if [[ $n == $unm && $pss == $p ]];
+	then
+		echo "Access Granted!"
+
+	else
+		echo "Un-Authorized User ..exiting"
+		sleep 3
+		kill -9 $PPID
+	fi
+}
+
+#Checks if the user is running the script for first time, it's not the proper way of doing this i guess but for now it works :P
+if [ -f "$pt" ];
 then
-echo "Access Granted!"
-
+	chk_crd
 else
-echo "Un-Authorized User... Exiting"
-sleep 3
-#kills the parent process id i.e., the app itself.
-kill -9 $PPID 
+	st_crd
 fi
 
 trap SIGINT
